+++ vpype_gscrib/excepts/__init__.py
@@ -7,10 +7,12 @@ This module defines a collection of exceptions to handle errors related
 to G-code processing and generation.
 """
 
+from .plugin_errors import FileLoadError
 from .plugin_errors import ImageLoadError
 from .plugin_errors import VpypeGscribError
 
 __all__ = [
+    "FileLoadError",
     "ImageLoadError",
     "VpypeGscribError",
 ]