+++ vpype_gscrib/__init__.py
@@ -15,7 +15,7 @@ from . import processor
 from . import renderer
 from . import heightmaps
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 
 __all__ = [
     "config",