+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 1.0.1 (2025-04-18)
 
-* En G-code programs with M02 instead of M30.
-* Fix port type in direct write or socket.
+* End G-code programs with M02 instead of M30.
+* Fix port type in direct write to socket.
 * Upgrade to GScrib 1.0.1 (direct write fix).
 
 ## 1.0.0 (2025-04-16)